moved rendering folder
