+++ src/simularium/VisGeometry/rendering/CompositePass.ts
