+++ src/simularium/VisGeometry/rendering/ContourPass.ts
