+++ src/simularium/VisGeometry/rendering/GBufferPass.ts
