+++ src/simularium/VisGeometry/rendering/GaussianBlur.ts
