+++ src/simularium/VisGeometry/rendering/HitTestHelper.ts
