+++ src/simularium/VisGeometry/rendering/InstancedFiber.ts
