+++ src/simularium/VisGeometry/rendering/InstancedFiberShader.ts
