+++ src/simularium/VisGeometry/rendering/InstancedMesh.ts
