+++ src/simularium/VisGeometry/rendering/InstancedMeshShader.ts
