+++ src/simularium/VisGeometry/rendering/MultipassMaterials.ts
