+++ src/simularium/VisGeometry/rendering/RenderToBuffer.ts
