+++ src/simularium/VisGeometry/rendering/SSAO.ts
