+++ src/simularium/VisGeometry/rendering/SimulariumRenderer.ts
